TextLinesStore: Vector -> std::vector

+++ src/colorer/LineSource.h
@@ -19,14 +19,14 @@ public:
    * @param lno Line number, which will be used as
    * initial position of all subsequend parsing.
    */
-  virtual void startJob(int lno) {};
+  virtual void startJob(size_t lno) {};
 
   /**
    * Called by parser, when it has finished text parsing.
    * Could be used to cleanup objects, allocated by last
    * #getLine() call.
    */
-  virtual void endJob(int lno) {};
+  virtual void endJob(size_t lno) {};
 
   /**
    * Returns line of text with specified number.
@@ -38,7 +38,7 @@ public:
    * @param lno Requested line number
    * @return Unicode string, enwrapped into String class.
    */
-  virtual String* getLine(int lno) = 0;
+  virtual String* getLine(size_t lno) = 0;
 protected:
   LineSource() {};
   virtual ~LineSource() {};
